src/qa_app/webapp/requirements.txt
streamlit
st-annotated-text
nltk
pandas
torch
transformers
whoosh